Add Gateway API CRDs to the repo with kustomize (#2011)
Problem: We don't have a way to tie a version of the Gateway API to our release (or edge). There's also not a single source of truth for the API version.

Solution: Use kustomize to add a specific version of the CRDs to our codebase. This way we only have to manage the version in one place and we can tie it to our releases.