Increase timeout for tracing test
Problem: The tracing test can be flaky and traces occasionally don't show up in the collector in a timely manner.

Solution: Remove the test retry, and replace with a longer timeout for sending requests and checking for traces. Unable to produce the issue locally, but seen in pipelines, so hoping that this helps.