Add custom annotations into GatewayClass and NginxGateway manifests (#1993)
Problem: For deploying our service we use ArgoCD. For this, we have a Helm Chart that contains multiple resources that are using CRDs, that are not managed by this helm chart. When trying to sync this application, it fails if at least one CRD is not installed in the cluster. This is expected.

Solution: To work around this problem, we want to use SkipDryRunOnMissingResource option as annotation for GatewayClass and NginxGateway objects from the nginx-gateway-fabric chart