Refacter code to use new ObjectType
Problem: We currently use the variable objType
to refer to a client.Object's type (skeleton of
an object). We also use the variable obj to refer
to a full client.Object (object with fields filled
out). However, in many parts of the codebase these
two variables are used closely together and are
both of type client.Object which can be a little
confusing.

Solution: I created a new type called ObjectType
and refactored the codebase to utilize it.