Replace TODO route condition with an Accepted/False condition (#2228)
Problem: When a route references a Gateway that is ignored, NGF adds the TODO condition:
"The condition for this has not been implemented yet: Gateway is ignored. "
This does not communicate to the user that the route is not accepted because the Gateway 
it references is ignored.

Solution: Change this condition to be Accepted/False/GatewayIgnored.