Add possibility to provide custom pod annotations (#2250)
Problem: Currently the helm chart does not provide an option to 
configure custom annotations for the nginx-gateway-fabric pod(s).
In our scenario, this is inconvenient because the pre-configured annotations 
Prometheus specific, but we are using Grafana Alloy to scrape metrics and Alloy 
is expecting different annotations. Of course there are also other scenarios where 
custom annotations can be useful, for instance to configure exceptions for Gatekeeper policies.

Solution: This Pull Request adds the possibility to configure your own 
custom annotations which will be added to the nginx-gateway-fabric pods.