Use custom NGF Debian image for tests (#2260)
Problem: We're using a specific version of debian to create the VM. This
image is deprecated. We also install all the dependencies needed every
time we create the VM.

Solution: Create a custom image with all of our dependencies and specify
the image-family as a parameter so we can always use the latest version.