Avoid cloning the repo/downloading dependencies (#2280)
Problem: The VM clones the repo every time it's created and downloads
the go dependencies. This is time consuming.

Solution: Clone the repo in the base image and run `go mod download` so
that most of the dependencies are already present. When the VM starts
and the repo is nginxinc it will just do a git fetch on the repo to sync
it.