Automate Zero Downtime Scale tests (#2259)
Problem: We want our NFR tests to be fully automated to save developer time for each release cycle, and have a repetitive way of running the tests.

Solution: Automate the zero downtime scaling test. No longer collecting logs as done previously, because error logs would be unreliable to collect for pods that are scaling down (we don't have persistence that we can easily use in the automation to gather historic logs). Ultimately we are still ensuring that traffic is flowing and status updates occur, which are the important pieces here.