Add options for chart version and output command
Problem: There's not a way to specify the version of the Chart and to
see the command used to install.

Solution: Add a parameter to specify the version of the Chart. When
installing from a registry if no version is specified the latest tag
will be installed. For now only the edge version will set the edge tag.
Also output the command used to install.