Fix graceful recovery tests (#2440)
Problem: Graceful recovery tests were failing in the pipeline

Solutions:
- Increase timeout when testing for passing/failing traffic after a container or pod restart. 
The timeout was set to 20 seconds, with a polling interval of 500ms, but the per-request 
timeout was 10 seconds. This means we were only retrying the request twice when a request 
timed out because it couldn't connect. Increasing the overall timeout to 60 seconds increased 
the number of retries and eliminated the errors.
- Remove the kubectl drain flag --delete-local-data since it no longer exists and
 replace it with --delete-emptydir-data.