Add SnippetsFilter API (#2667)
Problem: As a user of NGF that has a need for an enhancement in the project, I want the ability to customize the NGINX configuration for NGF, So that I can utilize features that have not yet been exposed via configuration, Or so that I can work around a problem that has not yet been solved in NGF, but can be fixed through NGINX configuration.

Solution: Add the new SnippetsFilter CRD to allow for injecting custom nginx configuration to a routing rule. Apply configuration of valid SnippetsFilters referenced in HTTPRoutes and GRPCRoutes to the appropriate contexts in the NGINX config. If the SnippetsFilter referenced is invalid (wrong group or kind), the routing rule is not configured. If the SnippetsFilter cannot be resolved, the routing rule is configured, but the route will return a 500.

---------

Co-authored-by: salonichf5 <146118978+salonichf5@users.noreply.github.com>
Co-authored-by: Kate Osborn <50597707+kate-osborn@users.noreply.github.com>