Fix failing unit test
Problem: Intermittently one of our unit tests was failing.

Solution: My hunch as to why it was failing is because of the parallel nature would occasionally cause an object to be in the wrong state for a specific test. So that object is now being defined for every test instead of globally.