Add new "Get started" document (#2700)
This commit adds a new "Get started" document for NGINX Gateway Fabric.
It re-orders the top level page so that it is second after "Overview" as
a means to get a reader started as quickly as possible, and is a
combination of the existing kind guide, a Helm installation path and the
cafe-example on GitHub.

As part of this work, the original kind guide has been deleted: the only
context lost is regarding port forwarding, which this new document
mentions not requiring. The README for the cafe-example could also be
stripped down as a result: the step "skipped" regards how to use a
different hostname.

Both of these could instead be added to a general network configuration
document.

Co-authored-by: Saylor Berman <s.berman@f5.com>