feat: Update docs configuration for new logo
In a prior commit, the Hugo theme for the documentation was updated.
With that update, the new logo for NGINX Gateway Fabric was made
available (Previously used for Service Mesh).

This commit makes the necessary configuration changes for the new logo
to be available: there is a paired pull request for the main
documentation website.