Improve usage reporting in automated tests (#2971)
Problem: Pipelines are failing intermittently due to r33 reporting errors.

Solutions:
- In the graceful recovery and reconfig tests, do not fail if there are errors
 in the NGINX logs. Instead, log the errors and continue. In the reconfig tests, 
all NGINX error logs are captured in the results file. Since the graceful 
recovery tests are part of the functional tests that run on every push to a PR 
and merge to main, it is not desirable to write the error logs to a file. 
Instead, if an unexpected error is seen in the NGINX logs, the test is skipped 
and the error is logged in the test output. Since we are sending traffic through 
NGINX during the graceful recovery tests, we will still catch any 
unrecoverable NGINX errors.

- Reporting for conformance tests is disabled.