Add guide for UpstreamSettingsPolicy (#2987)
Problem: As a user, I want an example of how to configure upstream 
settings for my application.

Solution: Add a how-to guide describing how to configure 
upstream zone size and keepalives for services. Also adds 
UpstreamSettingsPolicy to the table in the custom policies 
document and moves the custom policies document closer 
to the top of the overview page to increase visibility.